+++ kubejs/server_scripts/constants.js
@@ -1,0 +1,27 @@
+let tMetalData = {
+    'copper': 1080,
+    'bismuth_bronze': 985, 
+    'black_bronze': 1070, 
+    'bronze': 950, 
+    'wrought_iron' : 1535, 
+    'steel': 1540, 
+    'black_steel': 1540, 
+    'red_steel': 1540,
+    'blue_steel': 1485,
+}
+
+let hides = [['small', 300, 1], ['medium', 600, 2], ['large', 900, 3]]
+
+let furTypes = [
+    'black_bear',
+    'caribou', 
+    'cougar',
+    'crocodile', 
+    'direwolf', 
+    'grizzly_bear',
+    'lion', 
+    'panther', 
+    'polar_bear', 
+    'sabertooth', 
+    'tiger'
+]